requirements.txt
asgiref==3.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
zipp==3.8.0
